Improve logging for traceability of issue #1
It was looking like there was an infinite loop occurring in HttpParser
parse(). So the main way to observe this was to wait for it to occur
and then see if the new logging confirmed it. That occurred early on
so then additional logic was added to check if InputStream read() - on
line 55 - was hitting a 'no more data' condition. At the same time,
it provides a rough work around when it does occur. However, it needs
further work to be a real solution.