Refactor use of (and replace) HttpParser
This class had a mix of concerns, to the extent it ended up trying to
also look after handing of IO - when really it should've just parsed
a block of test representing a HTTP start-line and headers. IO really
should stay within the control of HttpForwarder as much as possible.

Doing so, then HttpForwarder can also be the one to check the basics
of incoming data - such as, there being none which is what triggered
this issue.

This addresses the issues in #1 through the throwing of an exception
in HttpForwarder.readStartlineAndHeaders() if there is no data found.

Unfortunately there is also a goo dnumber of other code that needs tidy
up in here. So a few not strictly related changes are also in here.

Issue: #1